Don't call dependencies from dependencies. Also move header calls to top as suggested
